movie-recommender-genai/requirements.txt
weaviate-client>=4.5.5
fastembed
torch
ipython

